std: get rid of `sys_common::io`
